+++ src/pages/MovieOverview/MovieOverviewPage.tsx
@@ -1,7 +1,7 @@
 import "./MovieOverviewPage.css"
 import MovieCard from "@/components/MovieCard/MovieCard";
 import { IMovie } from "@/interfaces/IMovie";
-import { getMoviesUnlimited } from "@/util/apiPaths";
+import { getMoviesUnlimited } from "@/util/apiUtils";
 import { useEffect, useState } from "react";
 
 const MovieOverviewPage = () => {
